Refactor passwordCheck Fn
Renamed passwordValidation, using the event listener "input", verifies
the user's input and gives immediate feedback when the password rule is
passed.

At the end of the function, confirmPassword is executed to ensure that
the confirm field value is the same as the password field value.

The reason for returning 1 is that when submit is executed, it means
that the validation is passed.