Refactor formValidation
Press submit to start detection, except for the password field, other
fields will show an error message if no value is entered, if there is
a value entered then validation will be performed, a successful
validation will return the number 1, update let countValid.

At the end of the function, it will run to verify if the countValid
is 5, if it is, it means all the validation passed.